services/task-api/pyproject.toml
[project]
name = "task-management-api"
version = "0.1.0"
description = "Bu Proje api yazmayi ve kod yazmayi ogrenmek amaciyla gelistirilen To/Do api projesidir."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt[crypto]>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]

[tool.ruff.lint]
# Kontrol Kurallari
# E : pycodestyle errors
# F : pyflakes
# I : isort (import duzenleme)
# UP : pyupgrade (eski Python sozdizimini yeniler)
select = ["E","F","I","UP"]

[tool.mypy]
# Kullanilacak Python Surumu
# python_version : "3.12"
#Kati kurallar icin
# strict : true
explicit_package_bases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
